o Merge upstream https://github.com/emkey1/ish-AOK/commit/8abbc2e8102ef3965edc607c72c67b0293e24e0e (Add debug button to reset iOS mounts) o General code cleanup

+++ app/AppDelegate.h
@@ -13,7 +13,7 @@
 - (void)exitApp;
 
 #if !ISH_LINUX
-+ (int)bootError;
++ (intptr_t)bootError;
 #endif
 
 + (void)maybePresentStartupMessageOnViewController:(UIViewController *)vc;
